### 2026-01-23 9d35c76
Restructure paper.md for JOSS: add State of the Field, Software Design, Research Impact Statement, AI Usage Disclosure sections
