### 2026-01-23 #9887 @mirzafarangi
I have updated the paper to include the newly required sections:

- **State of the Field** - Added comparison of existing meta-analysis tools
- **Software Design** - Renamed and expanded with explicit "build vs. contribute" justification
- **Research Impact Statement** - Added with user adoption, citation evidence, and educational integration metrics
- **AI Usage Disclosure** - Now a standalone section with detailed disclosure

Please regenerate the PDF and re-check the paper structure.
