### 2026-01-23 #9887 @mirzafarangi
Based on the similar papers identified, the following reviewers may be suitable for this submission (not tagging):

- tomfaulkenberry (reviewed statsExpressions - R statistics package)
- nistara (reviewed StatAid - R GUI for data analysis)
- adithirgis (reviewed StatAid - R GUI for data analysis)

These reviewers have experience with R-based statistical tools with graphical interfaces, which aligns well with Meta-Mar's scope.

I'm also happy to leave reviewer selection entirely to the editors.
