### 2026-01-23 6a1b6c0
Update paper: clarify development history since 2019, update contact email
