### 2026-01-23 #9887 @mirzafarangi
@arfon — Thank you for the clarification request.

Meta-Mar has been under active development since 2019. The current GitHub repository (March 2025) represents a complete architectural rewrite for version 4.0, migrating from a legacy R Shiny codebase to the new architecture.

Earlier versions were not maintained in public version control. However, Meta-Mar's prior existence is documented in peer-reviewed literature. Our 2020 BMC Psychiatry publication references Meta-Mar v1.1.0:

"...we carried out initial calculations using Cochrane RevMan 5 and then repeated the calculations using Meta-Mar (1.1.0), a free online meta-analysis service developed by the first author of this study." — Beheshti, A., Chavanon, M. L., & Christiansen, H. (2020). Emotion dysregulation in adults with attention deficit hyperactivity disorder: a meta-analysis. BMC Psychiatry, 20(1), 120. https://doi.org/10.1186/s12888-020-02535-8

I will update the paper to clarify: "Meta-Mar has been in development since 2019, with the current version (4.0) representing a complete architectural rewrite initiated in March 2025."

Would this clarification address your concern?
