### 2026-02-15 #9887 @arfon
@VincentAlcazer @IndrajeetPatil @pmcharrison :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _Meta-Mar: An AI-Integrated Web Platform for Meta-Analysis_

Based on your experience, and your previous publications in JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
