### 2026-02-16 #9887 @pmcharrison
hi @arfon, sorry, I don't feel qualified, I've never done a meta-analysis myself. To be honest none of those automatically identified JOSS papers look relevant to me. I'd recommend doing a separate literature search and find someone who has written about meta-analysis methods.
