### 2026-02-17 #9887 @arfon
@DrMattG @kylehamilton @praveenkumarpgiindia – 👋 would you be willing to review this submission for JOSS? The submission under consideration is _Meta-Mar: An AI-Integrated Web Platform for Meta-Analysis_

Based on your experience, and your listed expertise in our reviewer system, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
