### 2026-02-18 #10075 @editorialbot
[REVIEW]: Meta-Mar: An AI-Integrated Web Platform for Meta-Analysis
**Submitting author:** @mirzafarangi (<a href="http://orcid.org/0000-0002-6377-0710">Ashkan Beheshti</a>)
**Repository:** https://github.com/mirzafarangi/meta-mar0.4.0.2
**Branch with paper.md** (empty if default branch):
**Version:** v4.0.2
