### 2026-02-26 mirzafarangi/meta-mar0.4.0.2#1 @DrMattG
Review comment 1 - are there different streams of development?
https://github.com/openjournals/joss-reviews/issues/10075#issuecomment-3928052413

The functionality of the live Shiny application appears to differ from the version available in the public GitHub repository. This raises concerns about reproducibility and whether the archived software fully represents the functionality described in the manuscript.

JOSS requires that the reviewed repository contain all core functionality necessary to reproduce the results and behaviour of the software. If there are differences between the deployed app and the repository (e.g. additional features, alternative implementations, or private components), these should be clearly documented and, where possible, incorporated into the open-source version.
