### 2026-02-26 mirzafarangi/meta-mar0.4.0.2#2 @DrMattG
Review comment 2 - generic approaches to bias assessments can be problematic
https://github.com/openjournals/joss-reviews/issues/10075#issuecomment-3928052413

Many “publication bias” features bundled in generic meta-analysis software are best viewed as *exploratory* (funnel/contour plots) or *fragile* (Egger/Begg) rather than definitive diagnostics or corrections. Under realistic heterogeneity, funnel asymmetry and small-study effects can appear even when there is no selection/publication bias, producing misleadingly “significant” Egger/Begg results. Conversely, when selection bias is present, trim-and-fill’s symmetry assumptions often do not match the missingness mechanism and can over- or under-correct the pooled effect.

Fail-safe N is particularly problematic as presented as “quantifying robustness”: it answers “how many null studies would remove significance” under unrealistic assumptions (missing studies are null, and significance is the target).

This is more pronounced in disciplines like mine (ecology) but anywhere with high heterogenity (expected high heterogenity) it will be an issue.

Using off-the-shelf tools without engaging the brain is the real issue (and not your problem to "solve"). But I would recommend much more cautious language in the documentation and in the app itself.

Some of these may help:
https://doi.org/10.1186/s13643-015-0004-8
https://doi.org/10.1111/2041-210X.13724
https://doi.org/10.1097/MD.0000000000015987
https://doi.org/10.3844/ajassp.2012.1512.1517
