### 2026-02-27 mirzafarangi/meta-mar0.4.0.2#3 @DrMattG
Review comment 3 - evidence of collaboration
On my checklist (https://github.com/openjournals/joss-reviews/issues/10075#issuecomment-3928052413) I have this item - "Collaborative effort: Does the commit history show contributions from multiple developers and evidence of iterative refinement through community feedback?"

I dont see any evidence of collaborative effort on the repo (but I could be looking in the wrong places!)
