### 2026-03-04 mirzafarangi/meta-mar0.4.0.2#4 @DrMattG
Review comment 4 - Installation instructions
I would recommend putting in simpler instructions for users to install the code - you could, for example add git to the Requirements, you could add some instructions for those without git:

---
## No Git?
Download the repo as a ZIP from GitHub, unzip it, then in R:

`setwd("path/to/meta-mar")`
`shiny::runApp(".")`
---

You could also add instructions for RStudio users to clone the repository (by again downloading the zip... or using usethis::create_from_github())

[link to Joss-review](https://github.com/openjournals/joss-reviews/issues/10075#issuecomment-3987087212)
